moved port_lookup.hpp, deleted test
