+++ tools/scaleout/CMakeLists.txt
@@ -46,7 +46,6 @@ target_sources(
         BASE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}
         FILES
             board/board.hpp
-            board/port_lookup.hpp
             cabling_generator/cabling_generator.hpp
             cabling_generator/regen_descriptors.hpp
             connector/connector.hpp
