+++ tt_metal/fabric/physical_system_discovery.cpp
@@ -24,7 +24,7 @@
 #include "tt_metal/llrt/hal.hpp"
 #include "tt_metal/fabric/serialization/physical_system_descriptor_serialization.hpp"
 #include "tt_metal/fabric/fabric_host_utils.hpp"
-#include <board/port_lookup.hpp>
+#include "tt_metal/fabric/port_lookup.hpp"
 
 namespace tt::tt_metal {
